# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Gabinete Digital Cícero João
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/team:
    get:
      tags: [Admin]
      operationId: listTeam
      summary: List gabinete staff members (governance)
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      tags: [Admin]
      operationId: createTeamMember
      summary: Invite/create a staff member by email (governance)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Team member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/team/{id}:
    patch:
      tags: [Admin]
      operationId: updateTeamMember
      summary: Update a staff member's cargo, permissions or active state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberUpdate"
      responses:
        "200":
          description: Team member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Permissions:
      type: object
      description: "Capability toggles for gabinete staff"
      properties:
        canManageDemands:
          type: boolean
        canRespondDemands:
          type: boolean
        canManageProjects:
          type: boolean
        canManageNews:
          type: boolean
        canManageAgenda:
          type: boolean
        canManageAppointments:
          type: boolean
        canReleaseScheduleCards:
          type: boolean
        canManageStats:
          type: boolean
        canManageVoters:
          type: boolean
        canMessageVoters:
          type: boolean
        canManageGifts:
          type: boolean
        canManageTeam:
          type: boolean
...
    TeamMember:
      type: object
      required: [id, role, permissions, active, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
        cargo:
          type: ["string", "null"]
        permissions:
          $ref: "#/components/schemas/Permissions"
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    TeamMemberInput:
      type: object
      required: [email, cargo]
      properties:
        email:
          type: string
          format: email
        name:
          type: ["string", "null"]
        cargo:
          type: string
        permissions:
          $ref: "#/components/schemas/Permissions"
    TeamMemberUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        cargo:
          type: string
        permissions:
          $ref: "#/components/schemas/Permissions"
        active:
          type: boolean
...